MMPretrained/mmpretrain/src/Services/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
click==8.1.6
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
fastapi==0.100.0
fonttools==4.41.1
h11==0.14.0
httptools==0.6.0
idna==3.4
kiwisolver==1.4.4
matplotlib==3.7.2
numpy==1.25.1
opencv-python==4.8.0.74
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pydantic==2.0.3
pydantic_core==2.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
tzdata==2023.3
uvicorn==0.23.1
watchfiles==0.19.0
websockets==11.0.3
